pyproject.toml
[tool.poetry]
name = "ai-whisper"
version = "0.1.0"
description = ""
authors = ["LinRenJie <xoxosos666@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
openai-whisper = { git = "https://github.com/openai/whisper.git", rev = "v20231117" }
setuptools-rust = "^1.9.0"
numpy = "^1.26.4"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"